Moved query responses to the shared module
